# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lugendo — Travel agency management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitations for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: sendInvitations
      tags: [invitations]
      summary: Send invitations to travelers for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitations sent
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
  /trips/{tripId}/invitations/{invitationId}:
    patch:
      operationId: updateInvitation
      tags: [invitations]
      summary: Update invitation (segment assignment)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: invitationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationUpdateInput"
      responses:
        "200":
          description: Updated invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "404":
          description: Not found
    delete:
      operationId: deleteInvitation
      tags: [invitations]
      summary: Delete an invitation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: invitationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SegmentValue:
      type: string
      nullable: true
      enum: [basic, standard, premium, null]
...
    Invitation:
      type: object
      required: [id, tripId, email, status, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        email:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        segment:
          $ref: "#/components/schemas/SegmentValue"
        travelerId:
          type: ["integer", "null"]
        travelerName:
          type: ["string", "null"]
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
    InvitationInput:
      type: object
      required: [invitees]
      properties:
        invitees:
          type: array
          items:
            type: object
            required: [email]
            properties:
              email:
                type: string
              segment:
                $ref: "#/components/schemas/SegmentValue"
    InvitationUpdateInput:
      type: object
      properties:
        segment:
          $ref: "#/components/schemas/SegmentValue"

    # ─── TRAVELER FRONT DESK ─────────────────────────────────────────────────
...